Refactor: #36 - 기타 로직 리팩토링
- 패키지 경로 수정
- 인덴트 수정
- 불필요한 클래스 삭제
-
