+++ src/main/java/com/fourfourfour/damoa/api/member/repository/MemberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 
     boolean existsByEmail(String memberEmail);
 
@@ -13,5 +13,5 @@ public interface MemberRepository extends JpaRepository<Member, Long>, MemberRep
 
     Optional<Member> findByEmail(String email);
 
-    Member findMemberBySeq(Long seq);
+    Optional<Member> findBySeq(Long seq);
 }
