--- src/main/java/com/fourfourfour/damoa/api/member/repository/MemberRepositoryCustom.java
@@ -1,10 +1,0 @@
-package com.fourfourfour.damoa.api.member.repository;
-
-import com.fourfourfour.damoa.api.member.dto.res.ResMemberDto;
-
-import java.util.Optional;
-
-public interface MemberRepositoryCustom {
-
-    Optional<ResMemberDto> findResMemberDtoByEmail(String email);
-}
