+++ src/main/java/com/fourfourfour/damoa/api/notice/controller/NoticeController.java
@@ -2,9 +2,8 @@ package com.fourfourfour.damoa.api.notice.controller;
 
 import com.fourfourfour.damoa.api.notice.dto.req.ReqNoticeDto;
 import com.fourfourfour.damoa.api.notice.service.NoticeService;
-import com.fourfourfour.damoa.common.auth.PrincipalDetails;
 import com.fourfourfour.damoa.common.dto.response.BaseResponseDto;
-import com.fourfourfour.damoa.common.message.Message;
+import com.fourfourfour.damoa.common.util.AuthenticationUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -24,17 +23,17 @@ public class NoticeController {
 
     private final NoticeService noticeService;
 
+    private final AuthenticationUtil authenticationUtil;
+
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN')")
     @ResponseStatus(CREATED)
     @PostMapping
-    public BaseResponseDto noticeRegister(@Valid @RequestBody ReqNoticeDto reqNoticeDto, Authentication authentication) {
+    public BaseResponseDto<?> noticeRegister(@Valid @RequestBody ReqNoticeDto reqNoticeDto, Authentication authentication) {
 
-        Long memberSeq = ((PrincipalDetails) authentication.getDetails()).getMember().getSeq();
+        Long memberSeq = authenticationUtil.getMemberSeq(authentication);
 
         noticeService.register(reqNoticeDto, memberSeq);
 
-        return BaseResponseDto.builder()
-                .message(Message.REGISTER_NOTICE)
-                .build();
+        return BaseResponseDto.builder().build();
     }
 }
