+++ src/main/java/com/fourfourfour/damoa/api/notice/service/NoticeServiceImpl.java
@@ -5,6 +5,7 @@ import com.fourfourfour.damoa.api.member.repository.MemberRepository;
 import com.fourfourfour.damoa.api.notice.dto.req.ReqNoticeDto;
 import com.fourfourfour.damoa.api.notice.entity.Notice;
 import com.fourfourfour.damoa.api.notice.repository.NoticeRepository;
+import com.fourfourfour.damoa.common.constant.ErrorMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -25,7 +26,8 @@ public class NoticeServiceImpl implements NoticeService{
     @Transactional
     public Notice register(ReqNoticeDto reqNoticeDto, Long memberSeq) {
 
-        Member findMember = memberRepository.findMemberBySeq(memberSeq);
+        Member findMember = memberRepository.findBySeq(memberSeq)
+                .orElseThrow(() -> new IllegalArgumentException(ErrorMessage.NULL_MEMBER));
 
         Notice newNotice = Notice.builder()
                 .title(reqNoticeDto.getTitle())
