+++ src/main/java/com/fourfourfour/damoa/common/auth/PrincipalDetailsService.java
@@ -2,7 +2,7 @@ package com.fourfourfour.damoa.common.auth;
 
 import com.fourfourfour.damoa.api.member.entity.Member;
 import com.fourfourfour.damoa.api.member.repository.MemberRepository;
-import com.fourfourfour.damoa.common.message.ErrorMessage;
+import com.fourfourfour.damoa.common.constant.ErrorMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
