+++ src/main/java/com/fourfourfour/damoa/common/constant/Message.java
@@ -1,6 +1,5 @@
-package com.fourfourfour.damoa.common.message;
+package com.fourfourfour.damoa.common.constant;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Component;
