--- src/main/java/com/fourfourfour/damoa/common/dto/exception/ErrorSerializer.java
@@ -1,28 +1,0 @@
-package com.fourfourfour.damoa.common.dto.exception;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import org.springframework.boot.jackson.JsonComponent;
-import org.springframework.validation.Errors;
-
-import java.io.IOException;
-
-@JsonComponent
-public class ErrorSerializer extends JsonSerializer<Errors> {
-    @Override
-    public void serialize(Errors errors, JsonGenerator generator, SerializerProvider provider) throws IOException {
-        generator.writeStartArray();
-        errors.getFieldErrors().stream().forEach(e -> {
-            try {
-                generator.writeStartObject();
-                generator.writeStringField("field", e.getField());
-                generator.writeStringField("defaultMessage", e.getDefaultMessage());
-                generator.writeEndObject();
-            } catch (IOException e1) {
-                e1.printStackTrace();
-            }
-        });
-        generator.writeEndArray();
-    }
-}
